+++ packages/cater-build/plugins/favicon.js
@@ -1,5 +1,4 @@
 // Copyright Jon Williams 2017-2018. See LICENSE file.
-const Events = require('../src/app-events');
 const fs = require('fs');
 const loaderUtils = require('loader-utils');
 const path = require('path');
@@ -42,6 +41,7 @@ class Favicon {
     this.appRootPath = app.appRootPath;
     this.assetPath = path.join(app.appRootPath, ...FAVICON_ASSETS_DIRECTORY);
     this.mapping = {};
+    this.metadata = {};
 
     this.serverContext = new ServerContext();
     app.serverContexts.favicon = this.serverContext;
@@ -59,20 +59,38 @@ class Favicon {
   // Assigns the current mapping values to the server context
   assignServerContext() {
     this.serverContext.clear();
+
+    // Microsoft Tile Meta-Data
+    const browserConfig = this.mapping['browserconfig.xml'];
+    if (browserConfig && this.metadata['browserconfig.xml']) {
+      this.serverContext.addMeta({ name: 'msapplication-config', content: browserConfig });
+
+      const match = this.metadata['browserconfig.xml'].match(/<TileColor>([^<]*)<\/TileColor>/);
+      if (match) this.serverContext.addMeta({ name: 'msapplication-TileColor', content: match[1] });
+    }
+
+    // Generate Theme Meta-Data
+    let color = null;
+    const siteManifest = this.metadata['site.webmanifest'];
+    if (siteManifest) {
+      color = JSON.parse(siteManifest).theme_color;
+      if (color) this.serverContext.addMeta({ name: 'theme', content: color });
+    }
+
+    // Note that Firefox (at v58) won't interrogate the icon sizes. It takes
+    // the last icon that is specified. Hence, the order of having the 16x16
+    // first is important here.
     Object.entries({
-      'apple-touch-icon.png': { rel: 'apple-touch-icon', sizes: '180x180' },
-      'favicon-32x32.png': { rel: 'icon', sizes: '32x32' },
       'favicon-16x16.png': { rel: 'icon', sizes: '16x16' },
-      'site.webmanifest': { rel: 'manifest ' }
+      'favicon-32x32.png': { rel: 'icon', sizes: '32x32' },
+      'apple-touch-icon.png': { rel: 'apple-touch-icon', sizes: '180x180' },
+      'safari-pinned-tab.svg': { rel: 'mask-icon', color },
+      'site.webmanifest': { rel: 'manifest' }
     }).forEach(([k, v]) => {
       v.href = this.mapping[k];
       if (!v.href) return;
       this.serverContext.addLink(v);
     });
-    const browserConfig = this.mapping['browserconfig.xml'];
-    if (browserConfig) {
-      this.serverContext.addMeta({ name: 'msapplication-config', content: browserConfig });
-    }
   }
 
   // Called when the Cater App is being built
@@ -112,6 +130,7 @@ class Favicon {
         const filename = addHashToFilename(content, name);
         const location = path.join((app.assetHost || '') + app.publicPath, filename);
         this.mapping[name] = location;
+        this.metadata[name] = content;
         compilation.assets[filename] = new RawSource(content);
       });
 
@@ -162,12 +181,4 @@ class Favicon {
   }
 }
 
-function plugin(cater) {
-  const favicon = new Favicon(cater);
-  cater._favicon = favicon;
-
-  [Events.built, Events.compiling].forEach((v) => cater.on(v, favicon[v].bind(favicon)));
-  return favicon;
-}
-
-module.exports = plugin;
+module.exports = (app) => new Favicon(app);