+++ packages/cater-build/src/index.js
@@ -5,7 +5,6 @@ const DefaultConfig = require('./config-default.js');
 const errors = require('./app-errors');
 const Events = require('./app-events');
 const fs = require('fs-extra');
-const merge = require('deepmerge');
 const path = require('path');
 const Plugins = require('./plugins.js');
 const SideConfiguration = require('./app-side.js');
@@ -20,8 +19,9 @@ class BuildCater extends RuntimeCater {
     // Build has several server context that are combined in the handler.
     // This allows plugins to maintain their own server-context state.
     this.serverContexts = {
-      internal: this.serverContext
+      internal: this.serverContext // TODO
     };
+    this.serverContext = new ServerContext(); // TODO
 
     this.configuredPlugins = Plugins(this, config);
 
@@ -31,6 +31,7 @@ class BuildCater extends RuntimeCater {
     this.hotModuleReplacement = config.hotModuleReplacement;
     this.startOnError = config.startOnError;
     this.universalNames = config.universalNames;
+    this.useLogging = config.useLogging;
 
     // EVENT: I know we're already configuring, but this is for the benefit
     // of the plugins - which have only just been set up.
@@ -49,7 +50,7 @@ class BuildCater extends RuntimeCater {
 
   // Returns a build as a Promise object
   build() {
-    const currentBuild = new Build(this.buildPath);
+    const currentBuild = new Build(this.appRootPath, this.buildPath);
 
     return fs
       .remove(this.buildPath) // rm -rf ./build
@@ -130,25 +131,27 @@ class BuildCater extends RuntimeCater {
     return Middleware(this);
   }
 
-  // Returns all the "sides" of this application, as an array of string. This
-  // will typically return `['app', 'client', 'server', 'assets']`.
-  listSides() {
-    // Assets hardcoded. See https://github.com/clashbit/cater/issues/1
-    return Object.values(this.sides)
-      .map((s) => s.name)
-      .concat(this.universalNames)
-      .concat('assets');
-  }
-
   mergeServerContexts() {
     // TODO: Re-Writes the server context in place. But likely there is
     // a much better way of handling this. We have separate server context
     // so plugins like assets and favicon don't need to track individual
     // context changes.
+
     const contexts = Object.values(this.serverContexts);
-    contexts.push({});
-    Object.entries(new ServerContext(merge(...contexts))).forEach(([k, v]) => {
-      this.serverContext[k] = v;
+    if (contexts.length === 0) {
+      throw new Error(
+        'Strange. BuildCater.serverContexts does not have any entries. It should at least have the internal value'
+      );
+    }
+
+    Object.keys(this.serverContext).forEach((key) => {
+      const values = contexts.map((v) => v[key]);
+      if (Array.isArray(this.serverContext[key])) {
+        const value = values.reduce((prev, curr) => prev.concat(curr), []);
+        this.serverContext[key] = value;
+      } else {
+        [this.serverContext[key]] = values;
+      }
     });
   }
 
@@ -168,6 +171,11 @@ class BuildCater extends RuntimeCater {
     return this.handler().then((handler) => {
       Middleware.httpServer(handler, this.httpPort);
     });
+  }
+
+  triggerDeploy() {
+    this.emit(Events.deploying, this);
+    this.emit(Events.deployed, this);
   }
 
   triggerWebpackCompiled(side, stats) {